moved SearchREsults into separate file
